+++ src/main/java/net/brnbrd/delightful/common/block/AcornSackItem.java
@@ -1,0 +1,11 @@
+package net.brnbrd.delightful.common.block;
+
+import net.brnbrd.delightful.common.item.IConfigured;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.level.block.Block;
+
+public class AcornSackItem extends BlockItem implements IConfigured {
+	public AcornSackItem(Block block, Properties properties) {
+		super(block, properties);
+	}
+}