+++ src/main/java/net/brnbrd/delightful/common/item/food/AcornItem.java
@@ -1,0 +1,7 @@
+package net.brnbrd.delightful.common.item.food;
+
+public class AcornItem extends DConsumableItem {
+	public AcornItem(Properties properties) {
+		super(properties);
+	}
+}